server/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
aiosqlite==0.20.0
asyncpg==0.29.0
websockets==13.0
python-telegram-bot==21.5
httpx==0.27.2
python-dotenv==1.0.1
cryptography==43.0.1
rich==13.8.1
